Create HeaderButtons component
